indoor/pipeline/requirements.txt
anthropic>=0.39.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
Pillow>=10.3.0
python-dotenv>=1.0.1
shapely>=2.0.4
